--- a/src/govuk/components/back-link/_style.scss
+++ b/src/govuk/components/back-link/_style.scss
@@ -1,0 +1,61 @@
+@include govuk-exports("govuk/component/back-link") {
+
+  .govuk-back-link {
+    @include govuk-typography-responsive($size: 16);
+    @include govuk-link-common;
+    @include govuk-link-style-text;
+
+    display: inline-block;
+    position: relative;
+
+    margin-top: govuk-spacing(3);
+    margin-bottom: govuk-spacing(3);
+
+    // Allow space for the arrow
+    padding-left: 14px;
+  }
+
+  // Only add a custom underline if the component is linkable
+  .govuk-back-link[href] {
+    // Use border-bottom rather than text-decoration so that the arrow is
+    // underlined as well.
+    border-bottom: 1px solid govuk-colour("black");
+
+    // Underline is provided by a bottom border
+    text-decoration: none;
+
+    // When the back link is focused, hide the bottom link border as the
+    // focus styles has a bottom border.
+    &:focus {
+      border-bottom-color: transparent;
+    }
+  }
+
+  // Prepend left pointing arrow
+  .govuk-back-link:before {
+    @include govuk-shape-arrow($direction: left, $base: 10px, $height: 6px);
+
+    content: "";
+
+    // Vertically align with the parent element
+    position: absolute;
+
+    top: 0;
+    bottom: 0;
+    left: 0;
+
+    margin: auto;
+  }
+
+  @if $govuk-use-legacy-font {
+    // Begin adjustments for font baseline offset
+    // These should be removed when legacy font support is dropped
+    .govuk-back-link:before {
+      $offset: 1px;
+
+      top: $offset * -1;
+      bottom: $offset;
+    }
+  }
+
+}