--- a/src/govuk/components/character-count/_character-count.scss
+++ b/src/govuk/components/character-count/_character-count.scss
@@ -2,4 +2,8 @@
 @import "../../tools/all";
 @import "../../helpers/all";
 
+@import "../error-message/error-message";
+@import "../hint/hint";
+@import "../label/label";
+
 @import "./style";