--- a/src/govuk/components/character-count/_style.scss
+++ b/src/govuk/components/character-count/_style.scss
@@ -1,0 +1,27 @@
+@import "../error-message/error-message";
+@import "../hint/hint";
+@import "../label/label";
+
+@include govuk-exports("govuk/component/character-count") {
+  .govuk-character-count {
+    @include govuk-responsive-margin(6, "bottom");
+
+    .govuk-form-group,
+    .govuk-textarea {
+      margin-bottom: govuk-spacing(1);
+    }
+
+    .govuk-textarea--error {
+      padding: govuk-spacing(1) - 2; // Stop a "jump" when width of border changes
+    }
+  }
+
+  .govuk-character-count__message {
+    margin-top: 0;
+    margin-bottom: 0;
+  }
+
+  .govuk-character-count__message--disabled {
+    visibility: hidden;
+  }
+}