--- a/src/govuk/components/checkboxes/_style.scss
+++ b/src/govuk/components/checkboxes/_style.scss
@@ -1,8 +1,3 @@
-@import "../error-message/error-message";
-@import "../fieldset/fieldset";
-@import "../hint/hint";
-@import "../label/label";
-
 @include govuk-exports("govuk/component/checkboxes") {
 
   $govuk-touch-target-size: 44px;