--- a/src/govuk/components/date-input/_date-input.scss
+++ b/src/govuk/components/date-input/_date-input.scss
@@ -2,29 +2,4 @@
 @import "../../tools/all";
 @import "../../helpers/all";
 
-@import "../error-message/error-message";
-@import "../input/input";
-@import "../hint/hint";
-@import "../label/label";
-
-@include govuk-exports("govuk/component/date-input") {
-  .govuk-date-input {
-    @include govuk-clearfix;
-    // font-size: 0 removes whitespace caused by inline-block
-    font-size: 0;
-  }
-
-  .govuk-date-input__item {
-    display: inline-block;
-    margin-right: govuk-spacing(4);
-    margin-bottom: 0;
-  }
-
-  .govuk-date-input__label {
-    display: block;
-  }
-
-  .govuk-date-input__input {
-    margin-bottom: 0;
-  }
-}
+@import "./style";