--- a/src/govuk/components/date-input/_date-input.scss
+++ b/src/govuk/components/date-input/_date-input.scss
@@ -2,4 +2,9 @@
 @import "../../tools/all";
 @import "../../helpers/all";
 
+@import "../error-message/error-message";
+@import "../input/input";
+@import "../hint/hint";
+@import "../label/label";
+
 @import "./style";