--- a/src/govuk/components/date-input/_style.scss
+++ b/src/govuk/components/date-input/_style.scss
@@ -1,8 +1,3 @@
-@import "../error-message/error-message";
-@import "../input/input";
-@import "../hint/hint";
-@import "../label/label";
-
 @include govuk-exports("govuk/component/date-input") {
   .govuk-date-input {
     @include govuk-clearfix;