--- a/src/govuk/components/error-message/_error-message.scss
+++ b/src/govuk/components/error-message/_error-message.scss
@@ -2,14 +2,4 @@
 @import "../../tools/all";
 @import "../../helpers/all";
 
-@include govuk-exports("govuk/component/error-message") {
-  .govuk-error-message {
-    @include govuk-font($size: 19, $weight: bold);
-
-    display: block;
-    margin-bottom: govuk-spacing(3);
-    clear: both;
-
-    color: $govuk-error-colour;
-  }
-}
+@import "./style";