--- a/src/govuk/components/error-message/_style.scss
+++ b/src/govuk/components/error-message/_style.scss
@@ -1,0 +1,11 @@
+@include govuk-exports("govuk/component/error-message") {
+  .govuk-error-message {
+    @include govuk-font($size: 19, $weight: bold);
+
+    display: block;
+    margin-bottom: govuk-spacing(3);
+    clear: both;
+
+    color: $govuk-error-colour;
+  }
+}