--- a/src/govuk/components/error-summary/_error-summary.scss
+++ b/src/govuk/components/error-summary/_error-summary.scss
@@ -2,4 +2,6 @@
 @import "../../tools/all";
 @import "../../helpers/all";
 
+@import "../../core/lists";
+
 @import "./style";