--- a/src/govuk/components/error-summary/_style.scss
+++ b/src/govuk/components/error-summary/_style.scss
@@ -1,0 +1,55 @@
+@import "../../core/lists";
+
+@include govuk-exports("govuk/component/error-summary") {
+
+  .govuk-error-summary {
+    @include govuk-text-colour;
+    @include govuk-responsive-padding(4);
+    @include govuk-responsive-margin(8, "bottom");
+
+    border: $govuk-border-width solid $govuk-error-colour;
+
+    &:focus {
+      outline: $govuk-focus-width solid $govuk-focus-colour;
+    }
+  }
+
+  .govuk-error-summary__title {
+    @include govuk-font($size: 24, $weight: bold);
+
+    margin-top: 0;
+    @include govuk-responsive-margin(4, "bottom");
+  }
+
+  .govuk-error-summary__body {
+    @include govuk-font($size: 19);
+
+    p {
+      margin-top: 0;
+      @include govuk-responsive-margin(4, "bottom");
+    }
+  }
+
+  // Cross-component class - adjusts styling of list component
+  .govuk-error-summary__list {
+    margin-top: 0;
+    margin-bottom: 0;
+  }
+
+  .govuk-error-summary__list a {
+    @include govuk-typography-weight-bold;
+
+    // Override default link styling to use error colour
+    &:link,
+    &:visited,
+    &:hover,
+    &:active {
+      color: $govuk-error-colour;
+    }
+
+    &:focus {
+      @include govuk-focused-text;
+    }
+  }
+
+}