--- a/src/govuk/components/error-summary/_style.scss
+++ b/src/govuk/components/error-summary/_style.scss
@@ -1,5 +1,3 @@
-@import "../../core/lists";
-
 @include govuk-exports("govuk/component/error-summary") {
 
   .govuk-error-summary {