--- a/src/govuk/components/file-upload/_file-upload.scss
+++ b/src/govuk/components/file-upload/_file-upload.scss
@@ -2,4 +2,8 @@
 @import "../../tools/all";
 @import "../../helpers/all";
 
+@import "../error-message/error-message";
+@import "../hint/hint";
+@import "../label/label";
+
 @import "./style";