--- a/src/govuk/components/phase-banner/_phase-banner.scss
+++ b/src/govuk/components/phase-banner/_phase-banner.scss
@@ -2,4 +2,6 @@
 @import "../../tools/all";
 @import "../../helpers/all";
 
+@import "../tag/tag";
+
 @import "./style";