We currently import the settings, tools and helpers layers within each component so that users can import just a single component (e.g. `@import govuk/components/button/button`) and it'll import its own dependencies.

However, Ruby Sass in particular is slow to resolve these imports, and so when importing `govuk/all`, the fact that every component imports the dependencies again has a big impact on the compilation time.

By moving everything except the imports into a separate file (`_style.scss`) and including that in the original component .scss file, we can then import just the styles when importing all and shortcut all of the imports.

We can use the same logic to remove imports of dependent components (e.g. character count imports the styles for textarea).

Finally, we can remove two redundant imports of the typography helper, as we already import all helpers.

## Ruby Sass

**Before**: 30s
**After**: 7.8s

```
➜ sass --version
Ruby Sass 3.7.4
```

### Baseline (`master`)

```
➜ time sass -C src/govuk/all.scss > /dev/null
sass -C src/govuk/all.scss > /dev/null  30.02s user 0.99s system 98% cpu 31.614 total

➜ time sass -C src/govuk/all.scss > /dev/null
sass -C src/govuk/all.scss > /dev/null  30.00s user 0.96s system 98% cpu 31.363 total

➜ time sass -C src/govuk/all.scss > /dev/null
sass -C src/govuk/all.scss > /dev/null  29.29s user 0.91s system 98% cpu 30.524 total
```

### After changes (`sass-perf`)

```
➜ time sass -C src/govuk/all.scss > /dev/null
sass -C src/govuk/all.scss > /dev/null  7.80s user 0.35s system 98% cpu 8.304 total

➜ time sass -C src/govuk/all.scss > /dev/null
sass -C src/govuk/all.scss > /dev/null  7.89s user 0.35s system 98% cpu 8.321 total

➜ time sass -C src/govuk/all.scss > /dev/null
sass -C src/govuk/all.scss > /dev/null  7.68s user 0.32s system 99% cpu 8.049 total
```

## LibSass

**Before**: 0.2s
**After**: 0.18s

```
➜ sassc -v
sassc: 3.6.1
libsass: 3.6.3
sass2scss: 1.1.1
sass: 3.5
```

### Baseline (`master`)

```
➜ time sassc src/govuk/all.scss > /dev/null
sassc src/govuk/all.scss > /dev/null  0.20s user 0.02s system 98% cpu 0.224 total

➜ time sassc src/govuk/all.scss > /dev/null
sassc src/govuk/all.scss > /dev/null  0.20s user 0.02s system 98% cpu 0.224 total

➜ time sassc src/govuk/all.scss > /dev/null
sassc src/govuk/all.scss > /dev/null  0.20s user 0.02s system 99% cpu 0.224 total
```

### After changes (`sass-perf`)

```
➜ time sassc src/govuk/all.scss > /dev/null
sassc src/govuk/all.scss > /dev/null  0.18s user 0.03s system 87% cpu 0.242 total

➜ time sassc src/govuk/all.scss > /dev/null
sassc src/govuk/all.scss > /dev/null  0.17s user 0.02s system 93% cpu 0.198 total

➜ time sassc src/govuk/all.scss > /dev/null
sassc src/govuk/all.scss > /dev/null  0.16s user 0.02s system 98% cpu 0.183 total
```

## Dart Sass

**Before**: 0.5s
**After**: 0.25s

```
➜ /usr/local/Cellar/sass/1.23.7/bin/sass --version
1.23.7
```

### Baseline (`master`)

```
➜ time /usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null
/usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null  0.53s user 0.22s system 100% cpu 0.743 total

➜ time /usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null
/usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null  0.50s user 0.21s system 99% cpu 0.714 total

➜ time /usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null
/usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null  0.53s user 0.23s system 100% cpu 0.754 total
```

### After changes (`sass-perf`)

```
➜ time /usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null
/usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null  0.23s user 0.10s system 101% cpu 0.331 total

➜ time /usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null
/usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null  0.24s user 0.10s system 101% cpu 0.332 total

➜ time /usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null
/usr/local/Cellar/sass/1.23.7/bin/sass src/govuk/all.scss > /dev/null  0.24s user 0.10s system 101% cpu 0.340 total
```

Closes #1671