@nickcolley 2020-02-27
Apart from it being a bit weird it seems like a good way forward, nice thinking.

Think we could think about the following:

- what people should do to get a permanent custom build. Should we deprecate the current `button/button` import pattern entirely?
- documentation explaining why we have two files? I don't think we do this for Nunjucks perhaps so maybe this is a nice to have.
- seems like big improvements on ruby sass, but we know the issues are more to do Rails wrappers running ruby code regardless of the compiler, since even the libsass sassc-rails implementation seemed to have issues. Should we run this benchmark against something more real world?