@36degrees 2020-02-27
Just on the last point whilst I think about the others…

> * seems like big improvements on ruby sass, but we know the issues are more to do Rails wrappers running ruby code regardless of the compiler, since even the libsass sassc-rails implementation seemed to have issues. Should we run this benchmark against something more real world?

I think the next step might be to publish this as a pre-release and ask GOV.UK to do some benchmarking to understand if it improves things for them?