@36degrees 2020-02-27
> This might be a breaking change for any folks (low risk) that import only `@components/all`? I think it's a low impact breaking change but may need to go into 4.0.0

I guess we could add a single import of the three layers to `components/all` to mitigate that, if we wanted to? It'd reduce performance slightly again, but should still have a significant positive impact overall.