@36degrees 2020-02-27
> What should people do to get a performant custom build.
> Should we deprecate the current `button/button` import pattern entirely?
> Maybe not worth the trouble unless people run into it.

Good question. I think I'd expect people to start with a copy of `govuk/all.scss` and customise it to be more specific / remove the bits they don't need?

```scss
@import "settings/all";
@import "tools/all";
@import "helpers/all";

@import "core/all";
@import "objects/all";

@import "components/button/style";
@import "components/input/style";
@import "components/header/style";
@import "components/footer/style";

@import "utilities/all";
// Removed overrides as not using them
```

The use of `/style` does seem a bit strange, but maybe no more so than `button/button`…

Another option might be just to make this a proper breaking change, and remove the ability to import a single component without having to also import the `settings`, `tools` and `helpers` layers. I'm not sure what we'd do about dependent components though?

I guess we should also bear in mind that at some point in the next 12 - 18 months we'll probably want to look at moving to the `@uses` syntax anyway?