@nickcolley 2020-02-27
I think we should be taking more time to consider deprecation, some things we can't communicate deprecation that easily but Sass we often can do so pretty well with warnings.