@36degrees 2020-03-16
Having reviewed this as a team, we think that although it should be a non-breaking change for most users, it:

- increases the complexity of the codebase
- adds yet another way that users could import GOV.UK Frontend
- does nothing to improve compilation performance for users that are importing individual components, e.g. @import govuk/components/button/button
- would be a breaking change for anyone importing govuk/components/all without also importing the underlying layers (which we think would be an unusual way to import it, but should currently work)

On that basis, we now think that the best approach would be to remove the settings, tools and helpers imports from the components entirely, and update our documentation to tell users to import those layers if they are importing individual components.

There's more information on next steps in https://github.com/alphagov/govuk-frontend/issues/1671#issuecomment-599429977.